Getting Superservice id and key from config file

+++ scripts/initDockerLicode.sh
@@ -62,6 +62,17 @@ run_rabbitmq() {
 }
 
 create_superservice() {
+ if [ ! -z $CONFIG ]; then
+    echo [licode] Configuration file was passed.
+    NUVE_ID=`grep "config.nuve.superserviceID" $ROOT/licode_config.js`
+    NUVE_ID=`echo $NUVE_ID |  cut -d '=' -f 2`
+    NUVE_ID=`echo $NUVE_ID | cut -d ';' -f 1`
+
+    NUVE_KEY=`grep "config.nuve.superserviceKey" $ROOT/licode_config.js`
+    NUVE_KEY=`echo $NUVE_KEY |  cut -d '=' -f 2`
+    NUVE_KEY=`echo $NUVE_KEY | cut -d ';' -f 1`
+  fi
+
   dbURL=`grep "config.nuve.dataBaseURL" $SCRIPTS/licode_default.js`
 
   dbURL=`echo $dbURL| cut -d'"' -f 2`
@@ -112,12 +123,14 @@ create_superservice() {
 
   echo [licode] SuperService ID "$SERVID"
   echo [licode] SuperService KEY "$SERVKEY"
-  cd $BUILD_DIR
-  replacement=s/_auto_generated_ID_/${SERVID}/
-  sed $replacement $SCRIPTS/licode_default.js > $BUILD_DIR/licode_1.js
-  replacement=s/_auto_generated_KEY_/${SERVKEY}/
-  sed $replacement $BUILD_DIR/licode_1.js > $ROOT/licode_config.js
-  rm $BUILD_DIR/licode_1.js
+  if [ -z $CONFIG ]; then
+    cd $BUILD_DIR
+    replacement=s/_auto_generated_ID_/${SERVID}/
+    sed $replacement $SCRIPTS/licode_default.js > $BUILD_DIR/licode_1.js
+    replacement=s/_auto_generated_KEY_/${SERVKEY}/
+    sed $replacement $BUILD_DIR/licode_1.js > $ROOT/licode_config.js
+    rm $BUILD_DIR/licode_1.js
+  fi
 
 }
 run_nuve() {
@@ -163,61 +176,57 @@ nvm use
 #  run_rabbitmq
 #fi
 
-if [ ! -f "$ROOT"/licode_config.js ]; then
-    
-    #Select config file.
-    if [ ! -z "$CONFIG" ]; then
-        cp "$SCRIPTS"/"$CONFIG".js "$ROOT"/licode_config.js
-    else
-        cp "$SCRIPTS"/licode_default.js "$ROOT"/licode_config.js
-    fi
-    
-    
-    #Erizo Agent configuration
-    if [ ! -z "$ERIZO_AGENT_IP"  ]
-    then
-      echo "config.erizoAgent.publicIP = '$ERIZO_AGENT_IP';" >> /opt/licode/licode_config.js
-    fi
+#Select config file.
+if [ ! -z "$CONFIG" ]; then
+    cp "$SCRIPTS"/"$CONFIG".js "$ROOT"/licode_config.js
+else
+    cp "$SCRIPTS"/licode_default.js "$ROOT"/licode_config.js
+fi
 
-    echo "config.erizo.minport = '$MIN_PORT';" >> /opt/licode/licode_config.js
-    echo "config.erizo.maxport = '$MAX_PORT';" >> /opt/licode/licode_config.js
+#Erizo Agent configuration
+if [ ! -z "$ERIZO_AGENT_IP"  ]
+then
+  echo "config.erizoAgent.publicIP = '$ERIZO_AGENT_IP';" >> /opt/licode/licode_config.js
+fi
+
+echo "config.erizo.minport = '$MIN_PORT';" >> /opt/licode/licode_config.js
+echo "config.erizo.maxport = '$MAX_PORT';" >> /opt/licode/licode_config.js
 
 
-    if [ "$ERIZOAGENT" = "true" ] || [ "$ERIZOCONTROLLER" = "true" ] || [ "$NUVE" = "true" ] ; then
-      if [ ! -z "$NUVE_ID" ] && [ ! -z "$NUVE_KEY" ]; then 
-          echo "config.nuve.superserviceID = '$NUVE_ID';" >> /opt/licode/licode_config.js
-          echo "config.nuve.superserviceKey= '$NUVE_KEY';" >> /opt/licode/licode_config.js
-      fi
-      if [ ! -z "$RABBIT_URL" ]; then 
-          echo "config.rabbit.url = '$RABBIT_URL';" >> /opt/licode/licode_config.js
-      fi
-    fi
+if [ "$ERIZOAGENT" = "true" ] || [ "$ERIZOCONTROLLER" = "true" ] || [ "$NUVE" = "true" ] ; then
+  if [ ! -z "$NUVE_ID" ] && [ ! -z "$NUVE_KEY" ]; then 
+      echo "config.nuve.superserviceID = '$NUVE_ID';" >> /opt/licode/licode_config.js
+      echo "config.nuve.superserviceKey= '$NUVE_KEY';" >> /opt/licode/licode_config.js
+  fi
+  if [ ! -z "$RABBIT_URL" ]; then 
+      echo "config.rabbit.url = '$RABBIT_URL';" >> /opt/licode/licode_config.js
+  fi
+fi
 
-    #Erizo Controller Configuration
-    if [ ! -z "$ERIZO_CONTROLLER_IP" ]
-    then 
-      echo "config.erizoController.publicIP = '$ERIZO_CONTROLLER_IP';" >> /opt/licode/licode_config.js
-    fi
-    
-    if [ ! -z "$ERIZO_PORT" ]
-    then 
-      echo "config.erizoController.port = '$ERIZO_PORT';" >> /opt/licode/licode_config.js
-    fi
-    if [ ! -z "$ERIZO_SSL" ]
-    then 
-      echo "config.erizoController.ssl = '$ERIZO_SSL';" >> /opt/licode/licode_config.js
-    fi
-    if [ ! -z "$ERIZO_LISTEN_SSL" ]
-    then 
-      echo "config.erizoController.listen_ssl = '$ERIZO_LISTEN_SSL';" >> /opt/licode/licode_config.js
-    fi
-    if [ ! -z "$ERIZO_LISTEN_PORT" ]
-    then 
-      echo "config.erizoController.listen_port = '$ERIZO_LISTEN_PORT';" >> /opt/licode/licode_config.js
-    fi
-    if [ ! -z "$TURN_USERNAME" ] && [ ! -z "$TURN_CRED" ] && [ ! -z "$PUBLIC_IP" ] ; then
-      echo "config.erizoController.iceServers = [{\"url\":\"stun:'$PUBLIC_IP':3478\"}, {\"username\":\"'$TURN_USERNAME'\",\"url\":\"turn:'$PUBLIC_IP':3478\",\"credential\":\"'$TURN_CRED'\"}, {\"username\":\"'$TURN_USERNAME'\",\"url\":\"turn:'$PUBLIC_IP':5349\",\"credential\":\"'$TURN_CRED'\"}]" >> /opt/licode/licode_config.js
-    fi
+#Erizo Controller Configuration
+if [ ! -z "$ERIZO_CONTROLLER_IP" ]
+then 
+  echo "config.erizoController.publicIP = '$ERIZO_CONTROLLER_IP';" >> /opt/licode/licode_config.js
+fi
+
+if [ ! -z "$ERIZO_PORT" ]
+then 
+  echo "config.erizoController.port = '$ERIZO_PORT';" >> /opt/licode/licode_config.js
+fi
+if [ ! -z "$ERIZO_SSL" ]
+then 
+  echo "config.erizoController.ssl = '$ERIZO_SSL';" >> /opt/licode/licode_config.js
+fi
+if [ ! -z "$ERIZO_LISTEN_SSL" ]
+then 
+  echo "config.erizoController.listen_ssl = '$ERIZO_LISTEN_SSL';" >> /opt/licode/licode_config.js
+fi
+if [ ! -z "$ERIZO_LISTEN_PORT" ]
+then 
+  echo "config.erizoController.listen_port = '$ERIZO_LISTEN_PORT';" >> /opt/licode/licode_config.js
+fi
+if [ ! -z "$TURN_USERNAME" ] && [ ! -z "$TURN_CRED" ] && [ ! -z "$PUBLIC_IP" ] ; then
+  echo "config.erizoController.iceServers = [{\"url\":\"stun:'$PUBLIC_IP':3478\"}, {\"username\":\"'$TURN_USERNAME'\",\"url\":\"turn:'$PUBLIC_IP':3478\",\"credential\":\"'$TURN_CRED'\"}, {\"username\":\"'$TURN_USERNAME'\",\"url\":\"turn:'$PUBLIC_IP':5349\",\"credential\":\"'$TURN_CRED'\"}]" >> /opt/licode/licode_config.js
 fi
 
 if [ "$NUVE" = "true" ]; then
